Real-Time Voting System(Next.js)/backend/requirements.txt
Flask
flask-cors
kafka-python